Cholesky implemented (use own makefile)

+++ Software-Only/bus_master_if.h
@@ -22,6 +22,6 @@ class bus_master_if : virtual public sc_interface
     public:
         virtual void Request(unsigned int mst_id, unsigned int addr, unsigned int op, unsigned int len) = 0;
         virtual bool WaitForAcknowledge(unsigned int mst_id) = 0;
-        virtual void ReadData(unsigned int &data) = 0;
-        virtual void WriteData(unsigned int data) = 0;
+        virtual void ReadData(double &data) = 0;
+        virtual void WriteData(double data) = 0;
 };
